+++ days/day_14/part2_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func TestPart2Answer(t *testing.T) {
-	expected := 3278997609887
+	expected := int64(3278997609887)
 
 	input := inputs.Load(14)
 	out, err := main.Part2(context.Background(), input)
